refactor: rendering refinements
