+++ src/components/Container.jsx
@@ -4,7 +4,8 @@ import useMediaQuery from '@mui/material/useMediaQuery'
 
 import useGenerate from '@hooks/useGenerate'
 import useRefresh from '@hooks/useRefresh'
-import { useStatic } from '@hooks/useStore'
+import { useStatic, useStore } from '@hooks/useStore'
+import Utility from '@services/Utility'
 
 import Map from './Map'
 import ScanOnDemand from './layout/dialogs/scanner/ScanOnDemand'
@@ -17,6 +18,17 @@ import {
   ControlledTileLayer,
   ControlledZoomLayer,
 } from './Layers'
+
+/** @param {{ target: import('leaflet').Map, type: string }} */
+function setLocationZoom({ target: map }) {
+  const { lat, lng } = map.getCenter()
+  const zoom = map.getZoom()
+  useStore.setState({ location: [lat, lng], zoom })
+  useStatic.setState({
+    timeOfDay: Utility.timeCheck(lat, lng),
+  })
+  if (map.hasEventListeners('fetchdata')) map.fire('fetchdata')
+}
 
 const MAX_BOUNDS = [
   [-90, -210],
@@ -37,7 +49,18 @@ export default function Container({ serverSettings, params, location, zoom }) {
     <MapContainer
       tap={false}
       center={location}
-      ref={(ref) => useStatic.setState({ map: ref })}
+      ref={(ref) =>
+        useStatic.setState((prev) => {
+          if (ref) {
+            ref.attributionControl.setPrefix(
+              prev.config.map.attributionPrefix || '',
+            )
+            ref.on('moveend', setLocationZoom)
+            ref.on('zoomend', setLocationZoom)
+          }
+          return { map: ref }
+        })
+      }
       zoom={
         zoom < serverSettings.config.map.minZoom ||
         zoom > serverSettings.config.map.maxZoom
